issue #3 retrieve and execute launch4j
Add task to copy the launch4j binary and its dependency libs to gradle's
runtime folder. Call this binary instead of the configured
'launch4jCmd' if the boolean 'externalLaunch4j' is false.
This changes the default behavior.
Still missing component is the retrieval of the support binaries for
launch4j like ld and windres.