Clarify task usage
In the discussion in #9 the question arose which task a develop is
supposed to run to generate an executable with launch4j.

tl;dr: Execute the `launch4j` task

First of all there should be a clarification on the tasks and their
origin:
The `createExe` task was the previous task name and is kept for backward
compatibility. Therefore it *calls* - in gradle terms: *depends on* -
the other tasks `createExeWithJar` and `createExeWithBin`. The latter is
just the old renamed `createExe` task and calls the launch4j binary
anywhere on the `$PATH`. The property `externalLaunch4j` decides which
of the two tasks is called.
I've kept the `lauch4j` task for backward compatibility, too, though
it's only  purpose is to call all other tasks.
If you would like to create an executable you were supposed to run one
of the `launch4j` or `createExe` tasks and that has not changed. Anyway
I'm inclined to the view that the `launch4j` task is and will be the one
to execute.