fix the test case for issue #88
This commit adjusts the test cases for issue #88 to fail as in the provided example case. Though with launch4j version below 3.50 this is irrelevant as the `bundledJrePath` is simply used without checking the java version. With launch4j 3.50 this has to be fixed as it will search the provided jre paths.